+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Unreleased
 
 * `Node` class - WIP on `node-class` branch
+* add `adapters.mqtt_logger.MqttLogger` that forwards python logging to a mqtt topic
 * rename `bridges` to `adapters`. Also rename `MqttBridge` to `MqttAdapter`. This is to avoid confusion with
 stand-alone [mqtt-bridge](https://gitlab.com/roxautomation/components/mqtt-bridge)